backend/pyproject.toml
[project]
name = "technical-training-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django ~= 5.2",
    "djangorestframework ~= 3.16",
    "django-cors-headers ~= 4.7",
    "djangorestframework-simplejwt ~= 5.5",
    "psycopg[binary] ~= 3.2",
    "ruff ~= 0.12",
    "watchfiles ~= 1.1.0",
    # "watchdog[watchmedo] ~= 6.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]
